% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{gol_method_ruptures}
\alias{gol_method_ruptures}
\title{Rotture di metodo nella serie GOL}
\format{
Un \code{data.table} con 3 righe e le colonne:
\describe{
\item{data}{\code{IDate}, data convenzionale della rottura (2025-01-01).}
\item{evento}{Descrizione sintetica dell'evento.}
\item{scope}{Ambito interessato (quali temi / quali colonne).}
\item{riferimento}{Documento di riferimento.}
}
}
\source{
\code{dataset_long/README.md}, sezione "Cambiamenti di definizione
lungo la serie".
}
\usage{
gol_method_ruptures
}
\description{
Tre eventi documentati di discontinuita' metodologica nei dati GOL, tutti
collocati al passaggio dal formato pre-2025 (ANPAL/MLPS) al formato INAPP
Focus GOL del 2025. Pensato per essere passato come parametro \code{ruptures}
a \code{\link[=plot_timeline]{plot_timeline()}}.
}
\keyword{datasets}
