% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{gol_storia_caratteristiche}
\alias{gol_storia_caratteristiche}
\title{Storia lunga GOL: caratteristiche dei beneficiari 2022-2025}
\format{
Un \code{data.table} long con colonne: \verb{data_riferimento, fonte, regione, percorso, caratteristica, modalita, variabile, unita, valore, confidenza, era, rescan_severity}.
}
\usage{
gol_storia_caratteristiche
}
\description{
Serie regionale dei beneficiari per caratteristiche anagrafiche
(genere, classe eta', cittadinanza, durata disoccupazione).
}
\seealso{
\code{\link[=gol_storia_caratteristiche_series]{gol_storia_caratteristiche_series()}}.
}
\keyword{datasets}
