% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{gol_storia_esiti}
\alias{gol_storia_esiti}
\title{Storia lunga GOL: esiti 2022-2025}
\format{
Un \code{data.table} long con colonne: \verb{data_riferimento, fonte, regione, percorso, variabile, unita, valore, confidenza, tema_storico, era, rescan_severity}.
}
\usage{
gol_storia_esiti
}
\description{
Serie regionale degli esiti del programma: politiche attivate
(tema F), occupazione a 60/90/180 giorni (tema H), LEP e formazione
(INAPP tav 2.1), occupati totali e per percorso (INAPP tav 2.2).
}
\seealso{
\code{\link[=gol_storia_esiti_series]{gol_storia_esiti_series()}}.
}
\keyword{datasets}
