% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{gol_storia_volumi}
\alias{gol_storia_volumi}
\title{Storia lunga GOL: presi in carico 2022-2025}
\format{
Un \code{data.table} long con colonne: \verb{data_riferimento, fonte, regione, percorso, variabile, unita, valore, confidenza, era, rescan_severity}.
}
\usage{
gol_storia_volumi
}
\description{
Serie regionale dei volumi presi in carico dal Programma GOL, integrata
fra storico ANPAL/MLPS (2022-2024) e INAPP mensile (2024-06/2025-12).
}
\seealso{
\code{\link[=gol_storia_volumi_series]{gol_storia_volumi_series()}}, \code{\link[=plot_timeline]{plot_timeline()}}.
}
\keyword{datasets}
